pyproject.toml
[tool.poetry]
name = "ilmt"
version = "0.1.0"
description = ""
authors = ["IjatAyam <ijat191999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
python-decouple = "^3.6"
django-cors-headers = "^3.13.0"
django-debug-toolbar = "^3.4.0"
serializers = "^0.2.4"
requests = "^2.28.0"
googletrans = "4.0.0rc1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
